Adds dispatch feature (#69)
* Master -> dev (#44)


* Add CAD Display from Herald (#65)

* Restructured Index.html and added endpoint for herald to push dispatch info to

* Styled Dispatch alerts, set dispatch alert duration to three minutes, and fixed odds and ends

* add dev env

* add dev workflow

* match database fields

* add css file to index.html

* this is what happens when you copy and paste a line

* eslint

* work on dispatch styling

* more styling

* add dispatch time

* fix dispatch time text

* Style Changes for Headsup Dispatch (#66)

* Changed Styling to capitalize dispatch info by request of RPIA dev-team coordinator

* Capitalize the start of words and Uppercase the determinant in Dispatch Alerts

* Moved Dispatch style to style-dispatch.scss

* BugFix: Fix issues with dispatches that occur within 3 minutes of each other (#68)

* Changed Styling to capitalize dispatch info by request of RPIA dev-team coordinator

* Capitalize the start of words and Uppercase the determinant in Dispatch Alerts

* Moved Dispatch style to style-dispatch.scss

* Fixed issues with dispatches that occur within 3 minutes of each other

* Correct Logic Errors in ClearDispatch Timeout

* Added authentication token for herald dispatch

* 1.5.0

* Clean Code and Add Admin Authentication (#70)

* Spell checking and code adjustments

* Added token authentication to admin POST endpoints

* restructure admin authentication

* fix double-dispatch clearing time issue

* update build status badge

---------

+++ server.js
@@ -17,12 +17,13 @@ const pool = mariadb.createPool({
     user: process.env.DB_USER,
     password: process.env.DB_PASSWORD,
     connectionLimit: 10,
-    charset: 'utf8mb4'
+    charset: 'utf8mb4',
 });
 
 // Initialize express app
 const PORT = process.env.PORT || 8080;
 const WEBSITE_ACCESS_TOKEN = process.env.WEBSITE_ACCESS_TOKEN;
+const HERALD_TOKEN = process.env.HERALD_TOKEN;
 
 // app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
@@ -61,29 +62,49 @@ app.get('/notes', async (_, res) => {
 });
 
 app.post('/call/create', async (req, res) => {
-    await calls.createCall(pool, req.body);
-    const response_data = await calls.getTotalCalls(pool);
-    io.emit('calls', response_data);
-    res.send(response_data);
+    if (WEBSITE_ACCESS_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    }
+    else {
+        await calls.createCall(pool, req.body);
+        const response_data = await calls.getTotalCalls(pool);
+        io.emit('calls', response_data);
+        res.send(response_data);
+    }
 });
 
 app.post('/note/create', async (req, res) => {
-    const response_data = await notes.createNote(pool, req.body.note);
-    io.emit('notes', await notes.getNotes(pool));
-    res.send(response_data);
+    if (WEBSITE_ACCESS_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    }
+    else {
+        const response_data = await notes.createNote(pool, req.body.note);
+        io.emit('notes', await notes.getNotes(pool));
+        res.send(response_data);
+    }
 });
 
 app.post('/note/delete', async (req, res) => {
-    const response_data = await notes.deleteNote(pool, req.body.note);
-    io.emit('notes', await notes.getNotes(pool));
-    res.send(response_data);
+    if (WEBSITE_ACCESS_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    }
+    else {
+        const response_data = await notes.deleteNote(pool, req.body.note);
+        io.emit('notes', await notes.getNotes(pool));
+        res.send(response_data);
+    }
 });
 
 app.post('/mishap/create', async (req, res) => {
-    await mishap.createMishap(pool, req.body.mishap);
-    const response_data = await mishap.getTotalMishaps(pool);
-    io.emit('mishaps', response_data);
-    res.send(response_data);
+    if (WEBSITE_ACCESS_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    }
+    else {
+        await mishap.createMishap(pool, req.body.mishap);
+        const response_data = await mishap.getTotalMishaps(pool);
+        io.emit('mishaps', response_data);
+        res.send(response_data);
+    }
 });
 
 app.get('/mishap', async (_, res) => {
@@ -93,7 +114,17 @@ app.get('/mishap', async (_, res) => {
 
 app.post('/chores', (req, res) => {
     io.emit('chores', req.body);
-    res.send({success: true});
+    res.send({ success: true });
+});
+
+app.post('/dispatch', (req, res) => {
+    if (HERALD_TOKEN !== req.query.token) {
+        res.sendStatus(403);
+    } else {
+        console.log('received herald dispatch');
+        io.emit('dispatch', req.body);
+        res.send({ success: true });
+    }
 });
 
 io.on('connection', async () => {
